Use title for romanisations in parallel text
so that the translation parallel text is free of Chinese characters
(and those awful red brackets, which make the site too confusing
because they look like Wikipedia 'page-does-not-exist' links).
Replace these with a title attribute holding the original Chinese.
Optionally append the bracketed original Chinese explicitly
in annotations, but not in the translation parallel text.
---
This is a big commit. God save!